refactor more link_helpers methods into own files & rename methods
no code changes
