+++ happs/observation/zomes/fulfillment/code/src/fulfillment_requests.rs
@@ -18,8 +18,10 @@ use hdk_graph_helpers::{
         delete_record,
     },
     links::{
-        link_entries_bidir,
         get_links_and_load_entry_data,
+    },
+    local_indexes::{
+        create_direct_index,
     },
 };
 
@@ -79,7 +81,7 @@ fn handle_create_fulfillment(fulfillment: &CreateRequest) -> ZomeApiResult<Respo
     )?;
 
     // link entries in the local DNA
-    let _results = link_entries_bidir(
+    let _results = create_direct_index(
         fulfillment_address.as_ref(),
         fulfillment.get_fulfilled_by().as_ref(),
         FULFILLMENT_FULFILLEDBY_LINK_TYPE, FULFILLMENT_FULFILLEDBY_LINK_TAG,
